added ui data validation
